Split out MultiClassInspectorTest for reuse
